# docs/api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "CheckMyMole API"
  },
  "host" : "prod.api.checkmoleapp.demo-redisys.com",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/users/me/lesions/{lesion}/reports" : {
      "get" : {
        "tags" : [ "Reports" ],
        "summary" : "List all of the reports in the lesion",
        "parameters" : [ {
          "name" : "lesion",
          "in" : "path",
          "description" : "ID of the lesion",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "temp_urls",
          "in" : "query",
          "description" : "Should the photo URLs be signed?",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_answers",
          "in" : "query",
          "description" : "Should answers be included (as \"answers\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_questions",
          "in" : "query",
          "description" : "Should questions be included (as \"answers[]\" -> \"question\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "How many rows should be skipped",
          "required" : false,
          "type" : "integer",
          "default" : 0
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "How many rows should be returned",
          "required" : false,
          "type" : "integer",
          "default" : 50
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Report"
              }
            }
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      },
      "post" : {
        "tags" : [ "Reports" ],
        "summary" : "Create a new report",
        "parameters" : [ {
          "name" : "lesion",
          "in" : "path",
          "description" : "ID of the lesion",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Report to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Report"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Report"
            }
          },
          "400" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      }
    },
    "/users/me/lesions/{lesion}/reports/{id}" : {
      "put" : {
        "tags" : [ "Reports" ],
        "summary" : "Update an existing report",
        "parameters" : [ {
          "name" : "lesion",
          "in" : "path",
          "description" : "ID of the lesion",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the report to update",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Report to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Report"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Report"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied or invalid JSON input passed"
          },
          "404" : {
            "description" : "Report not found"
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      },
      "delete" : {
        "tags" : [ "Reports" ],
        "summary" : "Delete an existing report",
        "parameters" : [ {
          "name" : "lesion",
          "in" : "path",
          "description" : "ID of the lesion",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the report to delete",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Report"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Report not found"
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      }
    },
...
    "/reports" : {
      "get" : {
        "tags" : [ "Reports" ],
        "summary" : "List all of the reports in the system",
        "parameters" : [ {
          "name" : "request_id",
          "in" : "query",
          "description" : "Filter reports by this particular request_id",
          "required" : false,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "lesion_id",
          "in" : "query",
          "description" : "Filter reports by this particular lesion_id",
          "required" : false,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "temp_urls",
          "in" : "query",
          "description" : "Should photo URLs be signed?",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_lesions",
          "in" : "query",
          "description" : "Should lesions be included (as the \"lesion\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_body_parts",
          "in" : "query",
          "description" : "Should body_parts be included (as the \"body_part\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_answers",
          "in" : "query",
          "description" : "Should answers be included (as the \"answers\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_questions",
          "in" : "query",
          "description" : "Should questions be included (as the \"answers[].question\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "How many rows should be skipped",
          "required" : false,
          "type" : "integer",
          "default" : 0
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "How many rows should be returned",
          "required" : false,
          "type" : "integer",
          "default" : 50
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Report"
              }
            }
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "reports.read" ]
        } ]
      }
    },
    "/reports/{id}" : {
      "get" : {
        "tags" : [ "Reports" ],
        "summary" : "Get a single report by its ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the report to fetch",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "include_lesion",
          "in" : "query",
          "description" : "Should the lesion be included (as the \"lesion\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_body_part",
          "in" : "query",
          "description" : "Should the body_part be included (as the \"body_part\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_answers",
          "in" : "query",
          "description" : "Should answers be included (as the \"answers\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_questions",
          "in" : "query",
          "description" : "Should questions be included (as the \"answers[].question\" field)",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Report"
            }
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "reports.read" ]
        } ]
      }
    }
  },
  "securityDefinitions" : {
    "checkmoleapp_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://auth.checkmoleapp.demo-redisys.com/login",
      "flow" : "implicit",
      "scopes" : {
        "body-parts.write" : "create, update and delete body parts",
        "questions.write" : "create, update and delete questions",
        "owned.lesions.read" : "read lesions, reports and requests owned by the user",
        "owned.lesions.write" : "modify lesions, reports and requests owned by the user",
        "requests.read" : "read all requests in the system",
        "reports.read" : "read all reports in the system",
        "requests.respond" : "update all requests in the system",
        "lesions.read" : "read all lesions in the system"
      }
    }
  },
  "definitions" : {
...
    "Report" : {
      "type" : "object",
      "required" : [ "lesion_id", "photos" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "request_id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "ID of the request that the report belongs to"
        },
        "lesion_id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "ID of the lesion that this report is about"
        },
        "photos" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "description" : "URL of the photo"
          }
        },
        "status" : {
          "type" : "string",
          "description" : "Any value is accepted"
        },
        "consultation_result" : {
          "type" : "string"
        },
        "answers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Report_answers"
          }
        }
      },
      "example" : {
        "lesion_id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "consultation_result" : "consultation_result",
        "answers" : [ {
          "question" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "answer" : "{}"
        }, {
          "question" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "answer" : "{}"
        } ],
        "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "request_id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "photos" : [ "photos", "photos" ],
        "status" : "status"
      }
    },
...
    "Report_answers" : {
      "properties" : {
        "question" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "ID of the question"
        },
        "answer" : {
          "type" : "object",
          "description" : "JSON value with the answer",
          "properties" : { }
        }
      },
      "example" : {
        "question" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "answer" : "{}"
      }
    }
  }
}